requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
django-cors-headers>=4.3.0
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
cryptography>=41.0.0
reportlab>=4.0.0
Pillow>=10.0.0
